Change CPF/CNPJ to lowercase in SQL queries and interfaces
In `usuario.js`, change `CPF` and `CNPJ` to `Cpf` and `Cnpj` in the SQL queries and interfaces.

In `vaga.interface.ts`, add `Email` and `Telefone` properties to the `IVaga` interface.

In `vaga-list-empresa.component.spec.ts` and `trabalhador-inicial.component.spec.ts`, update the mock data for `vagas` to include `Email` and `Telefone` properties.

In `vaga-list.component.spec.ts`, update the mock data for `vagas` to include `Email` and `Telefone` properties.

In `detalhe-vaga.component.html`, update the code to display `vaga?.Telefone` and `vaga?.Email` instead of `usuario.telefone` and `usuario.email`.

In `detalhe-vaga.component.spec.ts`, update the mock data for `usuario` to include `Cnpj` property.

In `detalhe-vaga.component.ts`, import `filter` from `rxjs` and update the code to display an error message if the user is already assigned to the job.